requirements.txt
discord.py==2.4.0
aiohttp==3.10.5
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
playwright==1.47.0
